+++ crates/api/test.sh
@@ -17,3 +17,5 @@ set -ex
 
 cargo check --features=host
 cargo check --features=wasm --target=thumbv7em-none-eabi
+cargo fmt -- --check
+cargo clippy -- --deny=warnings